April 13th Update
The game is now in a playable state. It is fully working with game Logic on both sides. I need to add a way for the Server to reset itself once a single game is finish. The clients also currently need to add a way to reset there own boards for the next game.